@@ -1,7 +1,7 @@
 name: my_games_list
 description: "A new Flutter project."
 publish_to: "none"
-version: 0.1.0
+version: 1.0.0
 
 environment:
   sdk: ^3.8.1
@@ -9,19 +9,19 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  mobx: ^2.4.0
-  flutter_mobx: ^2.2.1+1
-  get_it: ^8.0.2
-  go_router: ^14.6.1
-  shared_preferences: ^2.3.3
-  webview_flutter: ^4.10.0
+  mobx: ^2.5.0
+  flutter_mobx: ^2.3.0
+  get_it: ^8.2.0
+  go_router: ^16.2.4
+  shared_preferences: ^2.5.3
+  webview_flutter: ^4.13.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^5.0.0
-  build_runner: ^2.4.13
-  mobx_codegen: ^2.7.0
+  flutter_lints: ^6.0.0
+  build_runner: ^2.9.0
+  mobx_codegen: ^2.7.4
 
 flutter:
   uses-material-design: true
